cleaner project tile
